Moved `@tryghost/mw-session-from-token` to services folder
fix https://linear.app/ghost/issue/ENG-2085/mw-session-from-token

- this commit moves the lib code and tests to the core folder so we can
  colocate code in one place rather than splitting it out across
  packages, which increases the cognitive load and overhead
- this might be classed as a weird place to put the middleware - I'd
  agree, but I'd rather put all the code in one place and then fix it
  afterwards
