+++ ghost/core/core/server/services/auth/session/index.js
@@ -1,6 +1,6 @@
 const adapterManager = require('../../adapter-manager');
 const createSessionService = require('@tryghost/session-service');
-const sessionFromToken = require('@tryghost/mw-session-from-token');
+const sessionFromToken = require('./session-from-token');
 const createSessionMiddleware = require('./middleware');
 const settingsCache = require('../../../../shared/settings-cache');
 const {GhostMailer} = require('../../mail');
