+++ ghost/core/core/server/services/auth/session/session-from-token.js
