+++ ghost/core/package.json
@@ -132,7 +132,6 @@
     "@tryghost/mw-api-version-mismatch": "0.0.0",
     "@tryghost/mw-cache-control": "0.0.0",
     "@tryghost/mw-error-handler": "0.0.0",
-    "@tryghost/mw-session-from-token": "0.0.0",
     "@tryghost/mw-version-match": "0.0.0",
     "@tryghost/mw-vhost": "0.0.0",
     "@tryghost/nodemailer": "0.3.45",
