--- ghost/mw-session-from-token/.eslintignore
@@ -1,1 +1,0 @@
-*.ts
