--- ghost/mw-session-from-token/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/session-from-token');
